+++ gala/coordinates/gd1.py
@@ -60,6 +60,14 @@ class GD1Koposov10(coord.BaseCoordinateFrame):
                                             coord.SphericalRepresentation)):
             self._data.lon.wrap_angle = self._default_wrap_angle
 
+    # TODO: remove this. This is a hack required as of astropy v3.1 in order
+    # to have the longitude components wrap at the desired angle
+    def represent_as(self, base, s='base', in_frame_units=False):
+        r = super().represent_as(base, s=s, in_frame_units=in_frame_units)
+        r.lon.wrap_angle = self._default_wrap_angle
+        return r
+    represent_as.__doc__ = coord.BaseCoordinateFrame.represent_as.__doc__
+    
 
 # Rotation matrix as defined in the Appendix of Koposov et al. (2010)
 R = np.array([[-0.4776303088, -0.1738432154, 0.8611897727],